requirements.txt
# ------------------ Основные зависимости ------------------
pandas>=1.5.0
scikit-learn>=1.2.0
pyyaml>=6.0
numpy>=1.24.0

# ------------------ DVC (версионирование данных) ------------------
dvc>=3.0.0

# ------------------ MLflow (трекинг экспериментов) ------------------
mlflow>=2.0.0

# ------------------ Feast (Feature Store) ------------------
feast>=0.37.0
psycopg-binary>=3.0.0
psycopg-pool>=3.0.0

# ------------------ Marimo (реактивные ноутбуки) ------------------
marimo>=0.1.0

# ------------------ Валидация YAML ------------------
yamllint>=1.30.0

# ------------------ Диаграммы ------------------

docker-compose-diagram>=0.1.0
graphviz>=0.20.0

# ------------------Jupyter (для демонстрации) ------------------
jupyter>=1.0.0
ipykernel>=6.0.0

# ------------------ Дополнительно ------------------
tqdm>=4.65.0
requests>=2.28.0
